Basic JSON:API support.
This PR adds some support for the JSON:API content-type. This change
will parse the "links" object from the top-level document.

links objects appearing deeper in the document are currently ignored.
I'm hoping this is a good first foundation. From this we might be able
to figure out what the next steps are for dealing with embedded items
and links objects appearing elsewhere in the document.