A rewrite of 'follower'.
Currently this object behaves the same as the old one, but it does more
of the heavy lifting the 'Resource' first did.

By moving some of this logic into this new object, it becomes a
possibility to alter the behavior of fetching after calling .follow().

Specifically, this opens the door to alter the behavior of how fetching
links from a resource is done via a fluent interface. Examples:

// Using HEAD requests
resource.follow().useHead();

// Immediately fetch the body of the linked resource
resource.follow().prefetch();

// Send a Prefer-Push header
resource.follow().preferPush();

// Make sure we're not using caches
resource.follow().forceFresh();

// Include additional headers
resource.follow().withHeaders({ 'X-Foo': 'Bar' });

For most of these examples additional plumbing is required, but this is
a step forward.

cc @mattbishop

Updates #141