Support the 'invalidates' link.
When a HTTP method returns a Link header that contains a
rel="invalidates", then it means that the HTTP request that was recently
performed has caused other resources to change.

For example, given a resource /article/ that has a list of articles, and
a resource /article/123, running a DELETE request on /article/123 will
also cause a change to the articles list.

If the response to DELETE included the following header:

    Link: </article>; rel="invalidates"

Then this can signal a client that the cache for the /article
collection is no longer valid.

The 'invalidates' link relationship is specified in this draft:

https://tools.ietf.org/html/draft-nottingham-linked-cache-inv