Action system rewrite.
It's greatly simplified but makes more assumptions about the structure
of an action.

Actions are now basically a simple serialization for a HTML form. The
supported formats all fit in this structure, but it's possible that in
the future this is no longer the case.

This makes it possible however to take an action object, and have all
the information available to render a full HTML form.

This is a BC break for actions, but given that actions were still
experimental, I think that's OK. This will be released eventually as 4.1.

@reda-alaoui what do you tihnk?