Support for dropdowns.
This adds support for the Hal Forms 'options' property. This lets us
create fields that allow a user to select from within a list of options.

In addition of just using an inline list, this also lets you select an
external endpoint to serve up the list of options, either as a simple
JSON document, or a link relationship from a hypermedia document.

Using a list of links from a hypermedia document is not supported by
HAL, even though the internal Ketting models do support it. To use this
feature you must either construct Ketting actions through code, or wait
until one of the Hypermedia documents supports this in the future.

Lastly, we're also doing some restructing again of all the fields.