Support for the "inv-by" link.
See:
https://datatracker.ietf.org/doc/html/draft-nottingham-linked-cache-inv-04#section-4

The 'inv-by' link lets a resource tie its cache status to the cache
status of another resource.

For example, if we have to resources:

```
/foo/1
/foo/1/comments
```

And `/foo/1/comments` returns a `inv-by` link that points to `/foo/1`,
it means that _if_ the cache of `/foo/1/comments` expires for any
reason, the cache for `/foo/1` will also expire.