Handling of dependencies in caches is moved to the Client.
This is a fix for #416.

All tests pass, so this should not be a breaking change. The Client is
getting more and more 'cache functions'.

I would have preferred these features to exist in the `src/cache`
directory, but doing so would be a BC break so we can only introduce
this in a major version.