Increase 'max listeners'.
When Ketting is used in the frontend, we want to encourage it being used
as a store for central state. If a piece of data is used in different
places, this will usually result in events being registered.

The default of 10 is quite low, and standard applications often start
emitting warnings for 'max listeners'.

This PR increases it to 500 for the Resource object. This is plenty of
room.