Only emit 'update' events after the entire state cache has been updated.
Currently Ketting has a race condition for updates with embedded items.
If items are embedded, we were sending out 'update' events even if the
entire request wasn't processed yet.

This can make it unpredictable to know whether a resource's parent or
child is updated.

This change processes an entire _embedded tree with sub-dependencies,
and only after the entire tree is cached it will start sending out
'update' events.

Now, when you receive an 'update' event the entire cache state is
consistent.