Upgrade from fetch-mw-oauth2 to @badgateway/oauth2-client
This new client has a lot more features. This PR creates a compatibility
layer for the old API so nothing changes yet.

In the future the goal is to remove the @badgateway/oauth2-client
dependency from ketting and only suggest it. The library itself has a
way to create a ketting-compatibile middleware.