Better hidden field support + required validation
If a HTML form has a hidden field, the user doesn't need to do anything
with these fields; they just get sent along with requests.

We want HAL forms to work the same way. If a HAL form has hidden fields,
we'll just adopt the values (unless overwritten). This has the
additional advantage that new hidden fields can be added without
changing clients.

This doesn't work for just hidden fields, any form field that has a
'value' will adopt this value unless overridden.

This PR also adds some support for validation. If a form field was
'required', and no value was specified for that field we'll now error.